Update the design of stored procedure helpers
